requirements.txt
pyaes==1.6.1
pyasn1==0.5.1
python-decouple==3.8
rsa==4.9
Telethon==1.34.0

reportlab~=4.1.0
pillow~=10.2.0
pdf2image~=1.17.0